rename constant.js to colorSetting.js
