+++ colorSetting.js
